COMPLETE refactor of modules
