+++ e2l-parser/src/e2gw_rpc_server/e2gw_rpc_server.rs
@@ -4,17 +4,18 @@ pub(crate) mod e2gw_rpc_server {
 
     // RPC
     use self::edge2_gateway_server::Edge2Gateway;
+    use std::sync::{Arc, Mutex};
     use tonic::{Request, Response, Status};
 
     // Include the generated proto file
     tonic::include_proto!("edge2gateway");
 
     pub struct Edge2GatewayServerStruct {
-        e2l_crypto: E2LCrypto,
+        e2l_crypto: Arc<Mutex<E2LCrypto>>,
     }
 
     impl Edge2GatewayServerStruct {
-        pub fn new(e2l_crypto: E2LCrypto) -> Self {
+        pub fn new(e2l_crypto: Arc<Mutex<E2LCrypto>>) -> Self {
             Self {
                 e2l_crypto: e2l_crypto,
             }
@@ -32,12 +33,14 @@ pub(crate) mod e2gw_rpc_server {
             let dev_addr = inner_request.dev_addr;
             let g_as_ed_compressed = inner_request.g_as_ed;
             let dev_public_key_compressed = inner_request.dev_public_key;
-            let g_gw_ed_compressed = self.e2l_crypto.handle_ed_pub_info(
+            let e2l_crypto = self.e2l_crypto.lock().expect("Could not lock");
+            let g_gw_ed_compressed = e2l_crypto.handle_ed_pub_info(
                 dev_eui,
                 dev_addr,
                 g_as_ed_compressed,
                 dev_public_key_compressed,
             );
+            std::mem::drop(e2l_crypto);
             // Check if the result is empty
             let reply: GwInfo;
             if g_gw_ed_compressed.is_empty() {
@@ -75,7 +78,9 @@ pub(crate) mod e2gw_rpc_server {
             let inner_request = request.into_inner();
             let dev_eui = inner_request.dev_eui;
             let _dev_addr = inner_request.dev_addr;
-            let response = self.e2l_crypto.remove_e2device(dev_eui);
+            let e2l_crypto = self.e2l_crypto.lock().expect("Could not lock");
+            let response = e2l_crypto.remove_e2device(dev_eui);
+            std::mem::drop(e2l_crypto);
             Ok(Response::new(response))
         }
 
@@ -85,7 +90,9 @@ pub(crate) mod e2gw_rpc_server {
         ) -> Result<Response<GwResponse>, Status> {
             let inner_request = request.into_inner();
             let device_list = inner_request.device_list;
-            let response = self.e2l_crypto.add_devices(device_list);
+            let e2l_crypto = self.e2l_crypto.lock().expect("Could not lock");
+            let response = e2l_crypto.add_devices(device_list);
+            std::mem::drop(e2l_crypto);
             Ok(Response::new(response))
         }
 
@@ -95,7 +102,9 @@ pub(crate) mod e2gw_rpc_server {
         ) -> Result<Response<GwResponse>, Status> {
             let inner_request = request.into_inner();
             let is_active = inner_request.is_active;
-            self.e2l_crypto.set_active(is_active);
+            let e2l_crypto = self.e2l_crypto.lock().expect("Could not lock");
+            e2l_crypto.set_active(is_active);
+            std::mem::drop(e2l_crypto);
             let response = GwResponse {
                 status_code: 0,
                 message: "Parameters Updated!".to_string(),
